requirements.txt
numpy>=1.21.0
pandas>=1.3.0
tushare>=1.2.89
akshare>=1.8.0
scikit-learn>=1.0.0
xgboost>=1.5.0
optuna>=2.10.0
pytz>=2021.1
pyyaml>=6.0
matplotlib>=3.5.0
streamlit>=1.8.0
sqlalchemy>=1.4.0
plotly>=5.5.0
fastapi>=0.78.0
uvicorn>=0.17.0
backtrader>=1.9.76.123
pytest>=7.0.0
PyQt6>=6.4.0
PyQt6-Qt6>=6.4.0
PyQt6-sip>=13.4.0 